# api/http/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Pushkin API
  version: 1.0.0
  description: Multi-tenant notification delivery service.
servers:
  - url: /api/v1
...
security:
  - TenantAPIKey: []
paths:
...
  /providers:
    get:
      tags: [Providers]
      summary: List providers
      operationId: listProviders
      responses:
        '200': { description: Tenant providers, content: { application/json: { schema: { $ref: '#/components/schemas/ListProvidersResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Providers]
      summary: Create a provider
      operationId: createProvider
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateProviderRequest' } } } }
      responses:
        '201': { description: Provider created, content: { application/json: { schema: { $ref: '#/components/schemas/CreateProviderResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /providers/{provider_id}:
    get:
      tags: [Providers]
      summary: Get a provider
      operationId: getProvider
      parameters: [{ $ref: '#/components/parameters/ProviderID' }]
      responses:
        '200': { description: Provider configuration without credentials, content: { application/json: { schema: { $ref: '#/components/schemas/ProviderResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    TenantAPIKey: { type: apiKey, in: header, name: X-Pushkin-API-Key }
    AdminKey: { type: apiKey, in: header, name: X-Pushkin-Admin-Key }
  parameters:
...
    ProviderID: { name: provider_id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Request is malformed or violates validation rules. }
    Unauthorized: { description: Authentication failed or the caller is not authorized. }
    NotFound: { description: The requested resource does not exist. }
...
    Unavailable: { description: A required dependency is temporarily unavailable. }
    InternalError: { description: An unexpected server error occurred. }
  schemas:
...
    CreateProviderRequest:
      type: object
      required: [type, credentials, rate_limit_qps, rate_limit_burst]
      properties:
        type: { type: string, enum: [fcm] }
        credentials: { type: string, description: Credential material; encrypted before persistence and never returned. }
        rate_limit_qps: { type: integer, minimum: 1 }
        rate_limit_burst: { type: integer, minimum: 1 }
    CreateProviderResponse:
      type: object
      required: [provider_id]
      properties: { provider_id: { type: string, format: uuid } }
    ProviderResponse:
      type: object
      required: [provider_id, type, status, rate_limit_qps, rate_limit_burst]
      properties:
        provider_id: { type: string, format: uuid }
        type: { type: string, enum: [fcm] }
        status: { type: string, enum: [active, disabled] }
        rate_limit_qps: { type: integer, minimum: 1 }
        rate_limit_burst: { type: integer, minimum: 1 }
    ListProvidersResponse:
      type: object
      required: [providers]
      properties:
        providers:
          type: array
          items: { $ref: '#/components/schemas/ProviderResponse' }
...